+++ src/Models/NameOrder.php
@@ -16,7 +16,7 @@ use Konekt\Enum\Enum;
 
 class NameOrder extends Enum implements NameOrderContract
 {
-    const __default = self::WESTERN;
+    const __DEFAULT = self::WESTERN;
 
     const WESTERN   = 'western';
     const EASTERN   = 'eastern';