+++ src/resources/manifest.php
@@ -2,5 +2,5 @@
 
 return [
     'name'    => 'Konekt Address Module',
-    'version' => '1.3.0'
+    'version' => '2.0-dev'
 ];